Termino parte de Api

+++ src/main/java/services/geodds/entities/Distancia.java
@@ -1,6 +1,6 @@
 package services.geodds.entities;
 
 public class Distancia {
-  public int valor;
+  public double valor;
   public String unidad;
 }
