+++ src/main/java/transporte/Transporte.java
@@ -6,7 +6,7 @@ import services.geodds.ServicioGeodds;
 import java.io.IOException;
 
 public interface Transporte {
-  public int distanciaEntre(PuntoUbicacion origen, PuntoUbicacion destino) throws IOException;
+  public double distanciaEntre(PuntoUbicacion origen, PuntoUbicacion destino) throws IOException;
 }
